# server/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/residents": {
      "post": {
        "operationId": "ResidentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResidentsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      },
      "get": {
        "operationId": "ResidentsController_fetch",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindAllResidentsDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/residents/{id}": {
      "put": {
        "operationId": "ResidentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResidentsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/residents/{status}": {
      "get": {
        "operationId": "ResidentsController_fetchByStatus",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindAllResidentsDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Brgy Lodiong Document Request APIs documents",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CIVIL_STATUS": {
        "type": "string",
        "enum": [
          "SINGLE",
          "MARRIED",
          "WIDOW"
        ]
      },
      "RESIDENT_STATUS": {
        "type": "string",
        "enum": [
          "PENDING",
          "REGISTERED",
          "DISAPPROVED"
        ]
      },
      "CreateResidentsDto": {
        "type": "object",
        "properties": {
          "civilStatus": {
            "$ref": "#/components/schemas/CIVIL_STATUS"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "RESIDENT"
            ]
          },
          "password": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RESIDENT_STATUS"
          },
          "disApprovedReason": {
            "type": "string"
          }
        },
        "required": [
          "civilStatus",
          "firstname",
          "lastname",
          "email",
          "contact",
          "address",
          "role",
          "password",
          "status"
        ]
      },
      "USER_ROLE": {
        "type": "string",
        "enum": [
          "ADMIN",
          "RESIDENT"
        ]
      },
      "FindAllResidentsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "firstname": {
            "type": "string"
          },
          "civilStatus": {
            "$ref": "#/components/schemas/CIVIL_STATUS"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RESIDENT_STATUS"
          },
          "disApprovedReason": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/USER_ROLE"
          }
        },
        "required": [
          "id",
          "firstname",
          "civilStatus",
          "lastname",
          "email",
          "contact",
          "address",
          "createdAt",
          "status",
          "role"
        ]
      },
...
    }
  }
}